refactor: move tr_port declaration to types.h
